MeanFieldToolkit.Build

MeanFieldToolkit.Build.BuildFromInteractions! — Method

BuildFromInteractions!(tbMFT::TightBindingMFT{T, R}, HoppingOrderLookup::Dict{Tuple, Matrix{ComplexF64}} ; refresh::Bool = true)
BuildFromInteractions!(bdgMFT::BdGMFT{T, R, S} , HoppingOrderLookup::Dict{Tuple, Matrix{ComplexF64}}, PairingOrderLookup::Dict{Tuple, Matrix{ComplexF64}} ; refresh::Bool = true)

Builds the MFT decomposed hoppings (and pairings) given a MFT object and the lookup tables for the expectation values of the different order parameters. If refresh is set to true, then the MFT bonds are deleted and rebuilt from scratch. Otherwise, the new bonds are appended to the existing bonds.