MeanFieldToolkit.InteractionConvert

MeanFieldToolkit.InteractionConvert.DensityToPartonCoupling — Method

DensityToPartonCoupling(Ui::Matrix{Float64}, Uj::Matrix{Float64}) --> Array{ComplexF64, 4}

Converts the density-density interaction matrix Ui and Uj into a parton coupling array U.

MeanFieldToolkit.InteractionConvert.SpinToPartonCoupling — Method

SpinToPartonCoupling(J::Matrix{Float64}, spin::Rational) --> Array{ComplexF64, 4}
SpinToPartonCoupling(J::Matrix{Float64}, SpinVec_A::Vector{Matrix{ComplexF64}}, SpinVec_B::Vector{Matrix{ComplexF64}}) --> Array{ComplexF64, 4}

Converts the exchange matrix J into a parton coupling array U using the spin matrices SpinVec_A and SpinVec_B. If spin is passed, then the spin matrices are generated using the SpinMats function.