MeanFieldToolkit.MFTPlot

MeanFieldToolkit.MFTPlot.PlotMFT! — Method

PlotMFT!(mft::TBMFTModel ; plot_labels::Vector{String} = getproperty.(mft.HoppingBlock.params, :label), plot_legend::Bool = true)
PlotMFT!(mft::BdGMFTModel ; plot_labels::Vector{String} = getproperty.(mft.HoppingBlock.params, :label), plot_legend::Bool = true)

Plots the order parameters of the given MFT object.

If plot_labels is passed, then only the order parameters with the given labels are plotted.

If plot_legend is passed, then the legend is shown in the plot.

MeanFieldToolkit.MFTPlot.PlotMFTEnergy! — Method

PlotMFTEnergy!(mft::T) where {T<:Union{TBMFTModel, BdGMFTModel}}

Plots the mean-field energy of the given MFT object.