MeanFieldToolkit.MFTIter

MeanFieldToolkit.MFTIter.DecomposeGr — Method

DecomposeGr(Gr::Array{Matrix{ComplexF64}, T}, param::Param{2, R}, uc::UnitCell{T}, bz::BZ)

Decomposes the Green's function into an order parameter. The order parameter is calculated by taking the trace decomposition of the Green's function on each bond the order parameter exists in, and then taking the mean of all the bond order parameters, following the pattern of the given order parameter.

MeanFieldToolkit.MFTIter.MFTIterator — Method

MFTIterator(Strengths::Vector{R}, tbMFT::TightBindingMFT{T, R}) --> Vector{R}
MFTIterator(Strengths::Vector{R}, bdgMFT::BdGMFT{T, R, R}) --> Vector{R}

Runs a single iteration of the mean-field theory on the given MFT object, and returns the new order parameters.